+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # bitrix24-php-sdk change log
 ## 2.0 DEV
 * remove all old code
+* migrate to Symfony HttpClient
 * add documentation webhook auth type
+* add OAuth 2.0 support
 
 ## 0.7.0 (11.07.2020)
 * add arguments in method `Bitrix24\Bizproc\Robot::add` for return results support