+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * migrate to Symfony HttpClient
 * add documentation webhook auth type
 * add OAuth 2.0 support
+* add Events support
 
 ## 0.7.0 (11.07.2020)
 * add arguments in method `Bitrix24\Bizproc\Robot::add` for return results support